+++ rxx/primitives.h
@@ -4,6 +4,7 @@
 #include "rxx/config.h"
 
 #include "rxx/access.h"
+#include "rxx/iter_traits.h"
 
 #include <concepts>
 #include <iterator>
@@ -13,12 +14,37 @@
 RXX_DEFAULT_NAMESPACE_BEGIN
 
 namespace ranges {
-using std::ranges::range_common_reference_t;
-using std::ranges::range_difference_t;
-using std::ranges::range_reference_t;
-using std::ranges::range_rvalue_reference_t;
-using std::ranges::range_size_t;
-using std::ranges::range_value_t;
+
+template <range R>
+using range_reference_t = iter_reference_t<iterator_t<R>>;
+
+template <range R>
+using range_rvalue_reference_t = iter_rvalue_reference_t<iterator_t<R>>;
+
+template <range R>
+using range_common_reference_t = iter_common_reference_t<iterator_t<R>>;
+
+template <range R>
+using const_iterator_t = const_iterator<iterator_t<R>>;
+
+template <range R>
+using const_sentinel_t = const_sentinel<sentinel_t<R>>;
+
+template <range R>
+using range_const_reference_t = iter_const_reference_t<iterator_t<R>>;
+
+template <range R>
+using range_const_rvalue_reference_t =
+    iter_const_rvalue_reference_t<iterator_t<R>>;
+
+template <sized_range R>
+using range_size_t = decltype(ranges::size(std::declval<R&>()));
+
+template <range R>
+using range_difference_t = iter_difference_t<iterator_t<R>>;
+
+template <range R>
+using range_value_t = iter_value_t<iterator_t<R>>;
 } // namespace ranges
 
 RXX_DEFAULT_NAMESPACE_END
